requirements.txt
aiogram>=3.0
asyncpg
python-dotenv